work

+++ src/helpers/PointToDegree.go
@@ -2,6 +2,6 @@ package helpers
 
 import "math"
 
-func PointToDegree(point [2]int) int {
-	return int(math.Atan2(float64(point[1]), float64(point[0])))
+func PointToDegree(point [2]int) float64 {
+	return math.Atan2(float64(point[1]), float64(point[0]))
 }
